[FIX] Fixed the betty style
