+++ find_path.c
@@ -1,8 +1,9 @@
 #include "shell.h"
 /**
  * *find_path - Find the PATH
- * @ac: Argument counter.
+ * @environ: Environ variable.
  *
+ * Return: string with path or NULL in failure.
  */
 
 char *find_path(char **environ)
@@ -22,7 +23,7 @@ char *find_path(char **environ)
 		{
 			path = _strdup(aux);
 			token = strtok(path, "=");
-			while(token != NULL)
+			while (token != NULL)
 			{
 				token = strtok(NULL, "=");
 				return (token);
@@ -35,7 +36,7 @@ char *find_path(char **environ)
 /**
  * print_env - Print environment variables.
  *
- * Return: 0 or -1 in failure.
+ * @environ: Environ variable.
  */
 
 void print_env(char **environ)
@@ -55,7 +56,7 @@ void print_env(char **environ)
 
 /**
  * _which - Find the directory of the command.
- * @cp_rec: Path received
+ * @p_rec: Path received
  * @first_arg: Command
  *
  */
@@ -68,7 +69,7 @@ void _which(char *p_rec, char *first_arg)
 	size = _strlen(first_arg);
 
 	path = _strdup(p_rec);
-	arg = _strdup(string_nconcat("/", first_arg, size -1));
+	arg = _strdup(string_nconcat("/", first_arg, size - 1));
 	size = _strlen(arg);
 
 	path_tok = strtok(path, ":");
